交换交友页面与广场页面
