+++ app/src/main/java/com/example/myapplication/Common.java
@@ -3,6 +3,6 @@ package com.example.myapplication;
 public class Common {
 
 
-    public static String URL = "http://172.27.21.201:8080";
+    public static String URL = "http://172.27.1.140:8080";
 
 }
